cleaning to ddd
