+++ src/i18n/uk-UA/frontend/config.php
