+++ src/tests/unit/components/ConfigurationTest.php
